+++ Changelog.md
@@ -22,9 +22,11 @@
 - [#227]: Split `SeError` from `DeError` in the `serialize` feature.
   Serialize functions and methods now return `SeError`.
 - [#810]: Return `std::io::Error` from `Writer` methods.
+- [#811]: Split `NamespaceError` from `Error`.
 
 [#227]: https://github.com/tafia/quick-xml/issues/227
 [#810]: https://github.com/tafia/quick-xml/pull/810
+[#811]: https://github.com/tafia/quick-xml/pull/811
 
 
 ## 0.36.2 -- 2024-09-20